backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
pymongo==4.6.0
motor==3.3.2
redis==5.0.1

# Data Collection
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
newsapi-python==0.2.7
python-dotenv==1.0.0
schedule==1.2.0

# NLP Libraries
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
gensim==4.3.2
nltk==3.8.1
scikit-learn==1.3.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring
loguru==0.7.2
